__pycache__/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
configparser==5.2.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.7
django-birthday==0.1.4
django-ckeditor==6.2.0
django-countries==7.2.1
django-crispy-forms==1.13.0

django-js-asset==1.2.2
django-jsignature==0.10
filelock==3.4.0

idna==3.3
Pillow==8.4.0
platformdirs==2.4.0
psycopg==3.0.5

python-decouple==3.5
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0

